Improved backend modularity reogranized/fixed code
