+++ server/data/Users.txt
@@ -1,5 +1,5 @@
 2eda5ab0-6a63-11ea-a4b4-ebd4268cc60a,true,Jim,hunter123,Tim@tim.com,212,4,10,2020-03-17T14:15:41.727Z,2020-03-17T00:00:00.000Z,
 2eda5ab0-6a63-11ea-a4b4-ebd4268cc61b,false,Tim,hunter1234,Jim@Jim.com,5234,5,12,2020-03-17T14:15:41.727Z,2020-03-17T00:00:00.000Z,
-948bf810-71e2-11ea-a8e0-73473bc162b1,false,dev,dev,dev@gmail.com,18540,0,0,2020-03-30T01:53:15.508Z,2020-03-29T17:27:35.698Z,
-9e9a98d0-71e6-11ea-b474-933b66303ca9,false,john,john,john@gmail.com,24860,0,0,2020-03-30T01:53:51.990Z,2020-03-29T17:56:30.557Z,
+948bf810-71e2-11ea-a8e0-73473bc162b1,false,dev,dev,dev@gmail.com,18080,0,0,2020-03-31T04:46:56.480Z,2020-03-29T17:27:35.698Z,
+9e9a98d0-71e6-11ea-b474-933b66303ca9,false,john,john,john@gmail.com,23900,0,0,2020-03-31T04:46:57.689Z,2020-03-29T17:56:30.557Z,
 8fd55dc0-71f6-11ea-a4ad-f1607ee8ca58,false,mack,mack,mack@gmail.com,81600,0,0,2020-03-30T01:53:19.483Z,2020-03-29T19:50:37.725Z,
